requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
lightning==2.4.0
lightning-utilities==0.11.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.2
packaging==24.1
pillow==10.4.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
pytorch-lightning==2.4.0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
torch==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.14.0
